Utilization Realtime Display

+++ agents_playground/core/simulation.py
@@ -3,17 +3,22 @@ Single file rewrite of coroutine based simulation.
 Prototyping the class design. Will break into modules if this pans out.
 """
 from __future__ import annotations
-from collections import OrderedDict
+
+from multiprocessing.connection import Connection
 import os
-from types import SimpleNamespace
-from typing import Callable, NamedTuple, Optional, Union
+import statistics
+from types import NoneType, SimpleNamespace
+from typing import NamedTuple, Optional, cast
 
 import dearpygui.dearpygui as dpg
+from agents_playground.core.constants import UPDATE_BUDGET
 
-from agents_playground.core.observe import Observable
-from agents_playground.core.sim_loop import SimLoop
+from agents_playground.core.observe import Observable, Observer
+from agents_playground.core.performance_monitor import PerformanceMetrics, PerformanceMonitor
+from agents_playground.core.sim_loop import SimLoop, SimLoopEvent, UTILITY_UTILIZATION_WINDOW
 from agents_playground.core.task_scheduler import TaskScheduler
 from agents_playground.core.callable_utils import CallableUtility
+from agents_playground.core.time_utilities import TimeUtilities
 from agents_playground.entities.entities_registry import ENTITIES_REGISTRY
 from agents_playground.renderers.color import BasicColors, Color
 from agents_playground.renderers.renderers_registry import RENDERERS_REGISTRY
@@ -43,13 +48,15 @@ class SimulationDefaults:
   PARENT_WINDOW_WIDTH_NOT_SET: int = 0
   PARENT_WINDOW_HEIGHT_NOT_SET: int = 0
   CANVAS_HEIGHT_BUFFER: int = 40
-  AGENT_STYLE_STROKE_THICKNESS: float = 1.0
-  AGENT_STYLE_STROKE_COLOR: Color = BasicColors.white.value
+  AGENT_STYLE_STROKE_THICKNESS: float =2.0
+  AGENT_STYLE_STROKE_COLOR: Color = BasicColors.black.value
   AGENT_STYLE_FILL_COLOR: Color = BasicColors.blue.value
   AGENT_STYLE_SIZE_WIDTH: int = 20
   AGENT_STYLE_SIZE_HEIGHT: int = 20
 
-class Simulation(Observable):
+calculate_task_utilization = lambda duration: round((duration/UPDATE_BUDGET) * 100) 
+
+class Simulation(Observable, Observer):
   """This class may potentially replace Simulation."""
   _primary_window_ref: Tag
 
@@ -64,30 +71,61 @@ class Simulation(Observable):
       sim_initial_state_dl_ref=dpg.generate_uuid(), 
       buttons={
         'sim': {
-          'run_sim_toggle_btn': dpg.generate_uuid()
+          'run_sim_toggle_btn': dpg.generate_uuid(),
+          'toggle_utility_plot_btn': dpg.generate_uuid()
         }
       }
     )
-    # self._layers: OrderedDict[Tag, RenderLayer] = OrderedDict()
+   
+    self.__performance_panel_id = dpg.generate_uuid() # TODO: Move to SimulationUIComponents
+    self.__fps_widget_id = dpg.generate_uuid() # TODO: Move to SimulationUIComponents
+    self.__time_running_widget_id = dpg.generate_uuid() # TODO: Move to SimulationUIComponents
+    self.__cpu_util_widget_id = dpg.generate_uuid() # TODO: Move to SimulationUIComponents
+    self.__cpu_util_plot_id = dpg.generate_uuid() # TODO: Move to SimulationUIComponents
+    self.__process_memory_used_widget_id = dpg.generate_uuid() # TODO: Move to SimulationUIComponents
+    self.__process_memory_used_plot_id = dpg.generate_uuid() # TODO: Move to SimulationUIComponents
+    self.__physical_memory_used_widget_id = dpg.generate_uuid() # TODO: Move to SimulationUIComponents
+    self.__physical_memory_used_plot_id = dpg.generate_uuid() # TODO: Move to SimulationUIComponents
+    self.__virtual_memory_used_widget_id = dpg.generate_uuid() # TODO: Move to SimulationUIComponents
+    self.__virtual_memory_used_plot_id = dpg.generate_uuid() # TODO: Move to SimulationUIComponents
+    self.__page_faults_widget_id = dpg.generate_uuid() # TODO: Move to SimulationUIComponents
+    self.__page_faults_plot_id = dpg.generate_uuid() # TODO: Move to SimulationUIComponents
+    self.__pageins_widget_id = dpg.generate_uuid() # TODO: Move to SimulationUIComponents
+    self.__pageins_plot_id = dpg.generate_uuid() # TODO: Move to SimulationUIComponents
+    
+    self.__utility_bar_plot_id = dpg.generate_uuid() # TODO: Move to SimulationUIComponents
+    self.__utility_percentiles_plot_id = dpg.generate_uuid() # TODO: Move to SimulationUIComponents
+    self.__time_spent_rendering_plot_id = dpg.generate_uuid() # TODO: Move to SimulationUIComponents
+    self.__time_spent_running_tasks_plot_id = dpg.generate_uuid() # TODO: Move to SimulationUIComponents
+    self._show_perf_panel: bool = False
+   
     self._title: str = "Set the Simulation Title"
     self._sim_description = 'Set the Simulation Description'
     self._sim_instructions = 'Set the Simulation Instructions'
     self._task_scheduler = TaskScheduler()
     self._pre_sim_task_scheduler = TaskScheduler()
-    self._sim_loop = SimLoop(scheduler = self._task_scheduler)
+    self._sim_loop: SimLoop | None = SimLoop(scheduler = self._task_scheduler)
+    self._sim_loop.attach(self)
+    self.__perf_monitor: PerformanceMonitor | None = PerformanceMonitor()
+    self.__perf_receive_pipe: Optional[Connection] = None
     self._scene_reader = scene_reader
 
   def __del__(self) -> None:
     logger.info('Simulation deleted.')
     
-
   @property
   def simulation_state(self) -> SimulationState:
-    return self._sim_loop.simulation_state
+    if self._sim_loop is not None:
+      return self._sim_loop.simulation_state
+    else:
+      raise Exception('SimLoop not initialized.')
 
   @simulation_state.setter
   def simulation_state(self, next_state: SimulationState) -> None:
-    self._sim_loop.simulation_state = next_state
+    if self._sim_loop is not None:
+      self._sim_loop.simulation_state = next_state
+    else:
+      raise Exception('SimLoop not initialized.')
 
   @property
   def primary_window(self) -> Tag:
@@ -112,6 +150,7 @@ class Simulation(Observable):
       height=parent_height, 
       on_close=self._handle_sim_closed):
       self._setup_menu_bar()
+      self._create_performance_panel(cast(int, parent_width))
       render()
 
   def _load_scene(self):
@@ -133,6 +172,7 @@ class Simulation(Observable):
     self._establish_context()
     self._run_pre_simulation_routines()
     self._initialize_layers()
+    self.__perf_receive_pipe = self.__perf_monitor.start(os.getpid())
     self._sim_loop.start(self._context)
 
   def _establish_context(self) -> None:
@@ -156,6 +196,7 @@ class Simulation(Observable):
     else:
       self._context.canvas.height = SimulationDefaults.PARENT_WINDOW_HEIGHT_NOT_SET
     
+    # TODO: This should all be in a defaults file. Not code. Should all be overridable in a scene file.
     self._context.agent_style.stroke_thickness = SimulationDefaults.AGENT_STYLE_STROKE_THICKNESS
     self._context.agent_style.stroke_color = SimulationDefaults.AGENT_STYLE_STROKE_COLOR
     self._context.agent_style.fill_color = SimulationDefaults.AGENT_STYLE_FILL_COLOR
@@ -173,26 +214,39 @@ class Simulation(Observable):
         with dpg.draw_layer(tag=rl.id):
           CallableUtility.invoke(rl.layer, {'context': self._context})
   
-  def _handle_sim_closed(self, sender, app_data, user_data):
+  def _handle_sim_closed(self):
     logger.info('Simulation: Closing the simulation.')
-    # 1. Kill the simulation thread.
-    self.simulation_state = SimulationState.ENDED
+    self.shutdown()
+
+    # 3. Notify the parent window that this simulation has been closed.
+    super().notify(SimulationEvents.WINDOW_CLOSED.value)
+
+  def shutdown(self) -> None:
+    logger.info('Simulation: Shutting down the simulation.')
+    # 1. Stop the simulation thread and task scheduler.
+    self.simulation_state = SimulationState.ENDED 
     self._task_scheduler.stop()
-    self._sim_loop.end()
-    self._sim_loop = None 
 
-    # 2. Remove dpg items that have bound callbacks to the sim instance.
-    dpg.delete_item(item = self._ui_components.sim_window_ref)
+    # 2. Stop the performance monitor process if it's going.
+    if self.__perf_monitor is not None:
+      self.__perf_monitor.stop()
+      self.__perf_monitor = None
 
-    # 4. Remove the reference to the simulations key components.
+    # 3. Kill the simulation thread.
+    if self._sim_loop is not None:
+      self._sim_loop.end()
+      self._sim_loop = None 
+
+    # 4. Remove dpg items that have bound callbacks to the sim instance.
+    if dpg.does_item_exist(item = self._ui_components.sim_window_ref):
+      dpg.delete_item(item = self._ui_components.sim_window_ref)
+
+    # 5. Remove the reference to the simulations key components.
     self._task_scheduler.purge()
     self._pre_sim_task_scheduler.purge()
 
     # 5. Purge the Context Object
     self._context.purge()
-
-    # 3. Notify the parent window that this simulation has been closed.
-    super().notify(SimulationEvents.WINDOW_CLOSED.value)
 
   def _setup_menu_bar(self):
     logger.info('Simulation: Setting up the menu bar.')
@@ -203,6 +257,104 @@ class Simulation(Observable):
         callback=self._run_sim_toggle_btn_clicked
       )
       self._setup_layers_menu()
+      # dpg.add_menu_item(label="Toggle Fullscreen", callback=lambda:dpg.toggle_viewport_fullscreen())
+      dpg.add_menu_item(label='utility', callback=self._toggle_utility_graph)
+
+  def _toggle_utility_graph(self) -> None:
+    self._show_perf_panel = not self._show_perf_panel
+    dpg.configure_item(self.__performance_panel_id, show=self._show_perf_panel)
+
+  def _create_performance_panel(self, plot_width: int) -> None:
+    TOOL_TIP_WIDTH = 350
+    with dpg.group(tag=self.__performance_panel_id, show=self._show_perf_panel):
+      with dpg.group(horizontal=True):
+        dpg.add_button(tag=self.__fps_widget_id, label="FPS", width=100, height=50)
+        with dpg.tooltip(parent=self.__fps_widget_id):
+          dpg.add_text("Frames Per Second averaged over 120 frames.", wrap=TOOL_TIP_WIDTH)
+        
+        dpg.add_button(tag=self.__time_running_widget_id, label="Uptime", width=120, height=50)
+        with dpg.tooltip(parent=self.__time_running_widget_id):
+          dpg.add_text("The amount of time the simulation has been running.", wrap=TOOL_TIP_WIDTH)
+          dpg.add_text("Format: dd:hh:mm:ss", wrap=TOOL_TIP_WIDTH)
+        
+        dpg.add_button(tag=self.__cpu_util_widget_id, label="CPU", width=120, height=50)
+        with dpg.tooltip(parent=self.__cpu_util_widget_id):
+          dpg.add_simple_plot(tag=self.__cpu_util_plot_id, height=40, width=TOOL_TIP_WIDTH)
+          dpg.add_text("The CPU utilization represented as a percentage.", wrap=TOOL_TIP_WIDTH)
+          dpg.add_text("Can be greater than 100% in the case of the simulation utilizing multiple cores.", wrap=TOOL_TIP_WIDTH)
+          dpg.add_text("The utilization calculation is performed by comparing the amount of idle time reported against the amount of clock time reported.", wrap=TOOL_TIP_WIDTH)
+      
+        dpg.add_button(tag=self.__process_memory_used_widget_id, label="Process Memory",  width=180, height=50)
+        with dpg.tooltip(parent=self.__process_memory_used_widget_id):
+          dpg.add_simple_plot(tag=self.__process_memory_used_plot_id, height=40, width=TOOL_TIP_WIDTH)
+          dpg.add_text("Unique Set Size (USS)", wrap=TOOL_TIP_WIDTH)
+          dpg.add_text("The amount of memory that would be freed if the process was terminated right now.", wrap=TOOL_TIP_WIDTH)
+        
+        dpg.add_button(tag=self.__physical_memory_used_widget_id, label="Memory",  width=180, height=50)
+        with dpg.tooltip(parent=self.__physical_memory_used_widget_id):
+          dpg.add_simple_plot(tag=self.__physical_memory_used_plot_id, height=40, width=TOOL_TIP_WIDTH)
+          dpg.add_text("Resident Set Size (RSS)", wrap=TOOL_TIP_WIDTH)
+          dpg.add_text("The non-swapped physical memory the simulation has used.", wrap=TOOL_TIP_WIDTH)
+        
+        dpg.add_button(tag=self.__virtual_memory_used_widget_id, label="Virtual",  width=180, height=50)
+        with dpg.tooltip(parent=self.__virtual_memory_used_widget_id):
+          dpg.add_simple_plot(tag=self.__virtual_memory_used_plot_id, height=40, width=TOOL_TIP_WIDTH)
+          dpg.add_text("Virtual Memory Size (VMS)", wrap=TOOL_TIP_WIDTH)
+          dpg.add_text("The total amount of virtual memory used by the process.", wrap=TOOL_TIP_WIDTH)
+        
+        dpg.add_button(tag=self.__page_faults_widget_id, label="Page Faults",  width=180, height=50)
+        with dpg.tooltip(parent=self.__page_faults_widget_id):
+          dpg.add_simple_plot(tag=self.__page_faults_plot_id, height=40, width=TOOL_TIP_WIDTH)
+          dpg.add_text("The number of page faults.", wrap=TOOL_TIP_WIDTH)
+          dpg.add_text("Page faults are requests for more memory.", wrap=TOOL_TIP_WIDTH)
+        
+        dpg.add_button(tag=self.__pageins_widget_id, label="Pageins",  width=180, height=50)
+        with dpg.tooltip(parent=self.__pageins_widget_id):
+          dpg.add_simple_plot(tag=self.__pageins_plot_id, height=40, width=TOOL_TIP_WIDTH)
+          dpg.add_text("The total number of requests for pages from a pager.", wrap=TOOL_TIP_WIDTH)
+          dpg.add_text("https://www.unix.com/man-page/osx/1/vm_stat", wrap=TOOL_TIP_WIDTH)
+
+      dpg.add_simple_plot(
+        tag=self.__utility_bar_plot_id, 
+        overlay='Frame Task Utility (%)',
+        height=40, 
+        width = plot_width,
+        min_scale = 0,
+        max_scale = 100
+      )
+      
+      dpg.add_simple_plot(
+        tag=self.__utility_percentiles_plot_id, 
+        overlay='Frame Task Utility % Percentiles',
+        height=40, 
+        width = plot_width,
+        histogram=True
+      )
+      
+      dpg.add_simple_plot(
+        tag=self.__time_spent_rendering_plot_id, 
+        overlay='Time Spent Rendering (ms)',
+        height=40, 
+        width = plot_width,
+      )
+      
+      dpg.add_simple_plot(
+        tag=self.__time_spent_running_tasks_plot_id, 
+        overlay='Time Spent Running Tasks (ms)',
+        height=40, 
+        width = plot_width,
+      )
+
+  def update(self, msg:str) -> None:
+    """Receives a notification message from an observable object."""
+    match msg:
+      case SimLoopEvent.UTILITY_SAMPLES_COLLECTED.value:   
+        if self._show_perf_panel:   
+          self._update_frame_performance_metrics()
+      case SimLoopEvent.TIME_TO_MONITOR_HARDWARE.value:
+        self._update_hardware_metrics()
+      case _:
+        logger.error(f'Simulation.update received unexpected message: {msg}')
 
   def _setup_layers_menu(self) -> None:
     logger.info('Simulation: Setting up layer\'s menu.')
@@ -270,3 +422,141 @@ class Simulation(Observable):
     logger.info('Simulation: Running pre-simulation tasks.')
     self._pre_sim_task_scheduler.consume()
     logger.info('Simulation: Done running pre-simulation tasks.')
+
+  def _update_frame_performance_metrics(self) -> None:
+    per_frame_samples = self._context.stats.per_frame_samples
+    task_samples                = per_frame_samples['running-tasks'].samples
+    task_utilization_samples    = list(map(calculate_task_utilization, task_samples))
+  
+    
+    avg_utility = round(statistics.fmean(task_utilization_samples), 2)
+    min_utility = min(task_utilization_samples)
+    max_utility = max(task_utilization_samples)
+    percentiles = statistics.quantiles(task_utilization_samples, n=100, method='inclusive')
+    
+    avg_rendering_time = round(statistics.fmean(per_frame_samples['rendering'].samples), 2)
+    min_rendering_time = round(min(per_frame_samples['rendering'].samples), 2)
+    max_rendering_time = round(max(per_frame_samples['rendering'].samples), 2)
+    
+    avg_task_time = round(statistics.fmean(task_utilization_samples), 2)
+    min_task_time = round(min(task_utilization_samples), 2)
+    max_task_time = round(max(task_utilization_samples), 2)
+    
+    dpg.set_value(
+      item = self.__utility_bar_plot_id, 
+      value = task_utilization_samples
+    )
+    
+    dpg.set_value(
+      item = self.__utility_percentiles_plot_id, 
+      value = percentiles
+    )
+    
+    dpg.set_value(
+      item = self.__time_spent_rendering_plot_id, 
+      value = per_frame_samples['rendering'].samples
+    )
+    
+    dpg.set_value(
+      item = self.__time_spent_running_tasks_plot_id, 
+      value = task_samples
+    )
+    
+    dpg.configure_item(
+      self.__utility_bar_plot_id, 
+      overlay=f'Frame Utility % (avg/min/max): {avg_utility}/{min_utility}/{max_utility}'
+    )
+    
+    dpg.configure_item(
+      self.__time_spent_rendering_plot_id, 
+      overlay=f'Time Spent Rendering (avg/min/max): {avg_rendering_time}/{min_rendering_time}/{max_rendering_time}'
+    )
+
+    dpg.configure_item(
+      self.__time_spent_running_tasks_plot_id, 
+      overlay=f'Time Spent Running Tasks (avg/min/max): {avg_task_time}/{min_task_time}/{max_task_time}'
+    )
+
+  def _update_hardware_metrics(self) -> None:
+    # Note: Not providing a value to Pipe.poll makes it return immediately.
+    try:
+      if self.__perf_receive_pipe is not None and \
+        self.__perf_receive_pipe.readable and \
+        self.__perf_receive_pipe.poll():
+        metrics: PerformanceMetrics = self.__perf_receive_pipe.recv()
+
+        dpg.configure_item(
+          self.__fps_widget_id, 
+          label = f"FPS: {dpg.get_frame_rate()}"
+        )
+        
+        uptime = TimeUtilities.display_seconds(metrics.sim_running_time)
+        dpg.configure_item(
+          self.__time_running_widget_id,
+          label = uptime
+        )
+
+        dpg.configure_item(
+          self.__cpu_util_widget_id,
+          label = f"CPU:{metrics.cpu_utilization.latest:.2f}"
+        )
+
+        dpg.set_value(
+          item = self.__cpu_util_plot_id, 
+          value = metrics.cpu_utilization.samples
+        )
+        
+        dpg.configure_item(
+          self.__process_memory_used_widget_id,
+          label = f"USS: {metrics.memory_unique_to_process.latest:.2f} MB"
+        )
+
+        dpg.set_value(
+          item = self.__process_memory_used_plot_id, 
+          value = metrics.memory_unique_to_process.samples
+        )
+        
+        dpg.configure_item(
+          self.__physical_memory_used_widget_id,
+          label = f"RSS: {metrics.non_swapped_physical_memory_used.latest:.2f} MB"
+        )
+
+        dpg.set_value(
+          item = self.__physical_memory_used_plot_id, 
+          value = metrics.non_swapped_physical_memory_used.samples
+        )
+        
+        dpg.configure_item(
+          self.__virtual_memory_used_widget_id,
+          label = f"VMS: {metrics.virtual_memory_used.latest:.2f} MB"
+        )
+
+        dpg.set_value(
+          item = self.__virtual_memory_used_plot_id, 
+          value = metrics.virtual_memory_used.samples
+        )
+        
+        
+        dpg.configure_item(
+          self.__page_faults_widget_id,
+          label = f"Page Faults: {metrics.page_faults.latest}"
+        )
+
+        dpg.set_value(
+          item = self.__page_faults_plot_id, 
+          value = metrics.page_faults.samples
+        )
+        
+        dpg.configure_item(
+          self.__pageins_widget_id,
+          label = f"Pageins: {metrics.pageins.latest}"
+        )
+
+        dpg.set_value(
+          item = self.__pageins_plot_id, 
+          value = metrics.pageins.samples
+        )
+    except EOFError as e:
+      print('The Performance Monitor send an EOFError. The process may have crashed.')
+      print(e)
+
